Enable basic ability to reload class with reload command
The basic idea here is that you can now modify a class file (e.g.
/etc/userctl/normal.class), followed by a ./userctl reload normal, and
userctld will then drop the old properties in favor of the new ones. If
it so happens that the new configuration file is no longer valid,
userctld will simply warn the user and keep trucking along with the old
properties.

Pretty cool!