Remove pointless code pretending a uid_t is not 32 bits
for all intensive purposes... there are some non-Linux systems where
uid_t is not 32 bits (and ancient Linux systems), but for our purposes
this is a safe assumption. In fact, systemd even assumes this (see
docs/UIDS-GIDS.md in systemd's git repo) and we rely on systemd, so
suffice to say that [probably buggy] code that dealt with the possibilty
of a non-32bit uid_t is not longer needed.