Create naive set-property command that sets transient resource controls
You can now do a ./userctl set-property classname CPUQuota=400% and
userctld will set the quotas of all the users to 400% as well as to any
new users that login in. That being said, currently these resource
controls are transient, meaning once the class or daemon is reloaded (or
restarted), the resource controls will not be applied.

The implementation for this is also pretty naive beyond the controls
being transient since the code just assumes all the resource controls
given to it are new and unique. In other words, it'll take some more
commits and work to get this to a decent state.