Reuse vector allocation as array if caller doesn't modify it
This is introduced in the pretend_vector_is_array() function, which
returns the internal vector allocation and adds a null terminator to the
end. Using this function makes it possible to prevent some allocations
of arrays that are returned using dbus. Those corresponding return dbus
functions don't modify the arrays, so we can avoid extra array
allocations which both makes the code cleaner, faster and safer.